refactor: 환경 변수 수정 및 폴더 구조 변경
