+++ src/pages/Everytime/Everytime.tsx
@@ -5,8 +5,8 @@ import { CheckCircleIcon, LoaderCircle } from "lucide-react";
 import { ClockAlert } from "lucide-react";
 import { useCallback, useEffect, useRef, useState } from "react";
 import type React from "react";
-import AlertBox from "../../ui/custom/alertbox";
-import NavigationButtons from "../../ui/custom/navigationButton";
+import AlertBox from "../../components/ui/custom/alertbox";
+import NavigationButtons from "../../components/ui/custom/navigationButton";
 import type { EverytimeValues } from "./Everytime.Schema";
 import EverytimeTimeTableLink from "./Everytime.TimeTableLink";
 import validateEverytime from "./Everytime.Validate";
