+++ src/pages/PersonalInfo/PersonalInfo.tsx
@@ -2,7 +2,7 @@ import { usePersonalInfoStore } from "@/stores/usePersonalInfoStore";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useEffect } from "react";
 import { FormProvider, useForm } from "react-hook-form";
-import NavigationButtons from "../../ui/custom/navigationButton";
+import NavigationButtons from "../../components/ui/custom/navigationButton";
 import {
   fetchPersonalInfoData,
   submitPersonalInfoData,
