+++ src/pages/PersonalInfo/field/studentPhoneNumber.tsx
@@ -1,6 +1,6 @@
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
-import { formatPhoneNumber } from "@/components/pages/PersonalInfo/PersonalInfo.helper";
+import { formatPhoneNumber } from "@/pages/PersonalInfo/PersonalInfo.helper";
 import { forwardRef, useCallback } from "react";
 import { useControllerField } from "../PersonalInfo.ControlledField";
