+++ src/pages/Survey/Survey.schema.ts
